Squished fixes
*Adb Colors
*Gapps Backup -cm
*update bufffer sizes
*add init.d and system init
*dont try and install recovery

ADB Logcat colored output.

This patch adds a new '-C' option to logcat so that the output is
colored with colors similar to the ones in DDMS.
Simply type "adb logcat -C" to use it.
Works well with bash in gnome-terminal.

Change-Id: Ie72199ecca0e04f2082c993f5763646bb09da61ea

do not install recovery on boot

set /system/etc/init.d/* permissions

ported from CM7

Change-Id: I279d78679f7d779d6cd6bbc6c834d94706937ae8

Update tcp buffersize settings

Soem of these values are from CAF, some of these are
from various shipping devices, some are educated
guesses based on the properties of each connection.

In general, the higher the bandwidth-delay product
of a network, the larger the buffer needs to be.

Change-Id: Idd8cd27727b5e66bd3a46be5f9853380acbda8ba

Include extra init file and fix dalvik cache

Change-Id: Ib53539b04ff37c1dde6492e717a25239ba916e37